server/python/requirements.txt
python-docx
python-pptx
openpyxl
